build: Unbreak YAML syntax in GH-hsoted runner; clean up dev requirements
The comment starting at column zero made this Action unparseable. But
the comment can also just be removed entirely by using the new
`dev-requirements` target, which after PR #31080/commit 2edbdcf7f4 now
also installs pip and edxapp.

I'm not exactly sure why `--src` is being specified here (is this a global
package installation, and therefore the editable deps would be installed
in the working dir?) but it can be specified even with pip-sync via
`PIP_SRC`. I also added it to the verify job since it should probably be
there as well.